Upgrade libraries

+++ client/src/app/todo/edit/edit.page.ts
@@ -13,7 +13,7 @@ import {NgForm} from '@angular/forms';
 })
 export class EditPage implements OnInit {
 
-  selectedTodo: Todo;
+  selectedTodo: Todo | undefined;
 
   constructor(private readonly route: ActivatedRoute,
               private readonly router: Router,
@@ -22,7 +22,7 @@ export class EditPage implements OnInit {
               private readonly todoService: TodoService) {
   }
 
-  async ngOnInit() {
+  async ngOnInit(): Promise<void> {
     const todoIdString = this.route.snapshot.paramMap.get('id');
     if (todoIdString) {
       this.selectedTodo = await this.todoService.getTodo(parseInt(todoIdString, 10));
@@ -36,7 +36,7 @@ export class EditPage implements OnInit {
     }
   }
 
-  async deleteTodo() {
+  async deleteTodo(): Promise<void> {
     if (this.selectedTodo) {
       const alert = await this.alertController.create({
         header: 'Delete Todo',
@@ -55,16 +55,20 @@ export class EditPage implements OnInit {
     }
   }
 
-  async save(form: NgForm) {
-    this.selectedTodo.subject = form.value.subject;
-    this.selectedTodo.description = form.value.description;
-    await this.todoService.save(this.selectedTodo);
-    await this.messagesService.showSuccessToast('Todo successfully saved', 1000);
-    await this.router.navigate(['/todos']);
+  async save(form: NgForm): Promise<void> {
+    if (this.selectedTodo) {
+      this.selectedTodo.subject = form.value.subject;
+      this.selectedTodo.description = form.value.description;
+      await this.todoService.save(this.selectedTodo);
+      await this.messagesService.showSuccessToast('Todo successfully saved', 1000);
+      await this.router.navigate(['/todos']);
+    }
   }
 
-  private async reallyDeleteTodo() {
-    await this.todoService.delete(this.selectedTodo);
-    await this.router.navigate(['/todos']);
+  private async reallyDeleteTodo(): Promise<void> {
+    if (this.selectedTodo) {
+      await this.todoService.delete(this.selectedTodo);
+      await this.router.navigate(['/todos']);
+    }
   }
 }
